display/requirements.txt
Flask
Flask-SocketIO
gunicorn
eventlet
openai
markdown
python-dotenv==1.0.1